requirements.txt
Flask==3.1.0
httpx==0.28.1
PySide6==6.9.1
qrcode==8.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
tzdata==2025.2
